added access control in all controllers

+++ app/Http/Controllers/AircraftsController.php
@@ -25,16 +25,27 @@ class AircraftsController extends Controller {
   }
 
   public function index(FleetList $fleet_list) {
+    $user = Auth::user();
+    if(!$fleet_list->is_accessible_by($user)){
+      return view('common.not_authorized');
+    }
     Session::put('backUrl', route('fleet_lists.aircrafts.index', $fleet_list));
     return view('fleet_lists.show', compact('fleet_list'));
   }
 
   public function create(FleetList $fleet_list) {
     $user = Auth::user();
+    if(!$fleet_list->is_accessible_by($user)){
+      return view('common.not_authorized');
+    }
     return view('aircrafts.create', compact('user', 'fleet_list'));
   }
 
   public function store(FleetList $fleet_list, Request $request) {
+    $user = Auth::user();
+    if(!$fleet_list->is_accessible_by($user)){
+      return redirect(Session::get('backUrl'))->with('message', 'Create error: not authorized');
+    }
     $this->validate($request, $this->rules);
     $input = Input::all();
     Aircraft::create($input);
@@ -42,16 +53,27 @@ class AircraftsController extends Controller {
   }
 
   public function show(FleetList $fleet_list, Aircraft $aircraft) {
+    $user = Auth::user();
+    if($fleet_list != $aircraft->fleet_list || !$fleet_list->is_accessible_by($user)){
+      return view('common.not_authorized');
+    }
     Session::put('backUrl', route('fleet_lists.aircrafts.show', array($fleet_list, $aircraft)));
     return view('aircrafts.show', compact('fleet_list', 'aircraft'));
   }
 
   public function edit(FleetList $fleet_list, Aircraft $aircraft) {
     $user = Auth::user();
+    if($fleet_list != $aircraft->fleet_list || !$fleet_list->is_accessible_by($user)){
+      return view('common.not_authorized');
+    }
     return view('aircrafts.edit', compact('user', 'fleet_list', 'aircraft'));
   }
 
   public function update(Request $request, FleetList $fleet_list, Aircraft $aircraft) {
+    $user = Auth::user();
+    if($fleet_list != $aircraft->fleet_list || !$fleet_list->is_accessible_by($user)){
+      return redirect(Session::get('backUrl'))->with('message', 'Update error: not authorized');
+    }
     $this->validate($request, $this->rules);
     $input = array_except(Input::all(), '_method');
     $aircraft->update($input);
@@ -60,6 +82,10 @@ class AircraftsController extends Controller {
 
 
   public function destroy(FleetList $fleet_list, Aircraft $aircraft) {
+    $user = Auth::user();
+    if($fleet_list != $aircraft->fleet_list || !$fleet_list->is_accessible_by($user)){
+      return redirect (Session::get('backUrl'))->with('message', 'Delete error: not authorized');
+    }
     $aircraft->delete();
     return Redirect::route('fleet_lists.show', $fleet_list->id)->with('message', 'Aircraft deleted.');
   }
